requirements.txt
streamlit>=1.37.0
google-genai>=0.3.0
groq>=0.11.0
tenacity>=8.2.3
pydantic>=2.7.0
python-dotenv>=1.0.1
requests>=2.32.3